Server: port cloud functions to TypeScript, Node v14
