+++ .gitignore
@@ -1,3 +1,4 @@
 *.bin
+build/
 data/
 node_modules/
