+++ .nvmrc
@@ -1,1 +1,1 @@
-lts/dubnium
+lts/fermium
