+++ server/.gcloudignore
@@ -13,4 +13,5 @@
 .git
 .gitignore
 
+build
 node_modules
