--- server/index.js
@@ -1,2 +1,0 @@
-exports.dataCapture = require('./dataCapture').dataCapture;
-exports.report = require('./report').report;
