+++ server/src/config.ts
@@ -1,0 +1,8 @@
+export const projectId = "molten-turbine-171801";
+export const datasetId = "meter";
+export const tableId = "readings";
+
+export const defaultTimezone = "US/Pacific";
+
+export const CUFT_PER_METER_TICK = 0.1; // meter ticks are 1/10 cubic foot
+export const CUFT_DECIMAL_PLACES = 7; // for rounding historic readings (where cuft was computed from gallons)
