+++ server/src/index.ts
@@ -1,0 +1,2 @@
+export {dataCapture} from './dataCapture';
+export {report} from './report';
