+++ server/tsconfig.json
@@ -1,0 +1,20 @@
+{
+  "compilerOptions": {
+    "lib": ["es2020"],
+    "module": "commonjs",
+    "target": "es2020",
+    "typeRoots": ["node_modules/@types"],
+
+    "strict": true,
+    "esModuleInterop": true,
+    "skipLibCheck": true,
+    "forceConsistentCasingInFileNames": true,
+    "noImplicitReturns": true,
+    "noUnusedLocals": true,
+
+    "outDir": "build",
+    "sourceMap": true,
+  },
+  "compileOnSave": true,
+  "include": ["src"]
+}
